Clean up indexes for searching facet filters
